requirements.txt
alembic==1.10.3
apispec==6.3.0
autopep8==2.0.2
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-smorest==0.41.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
Mako==1.2.4
MarkupSafe==2.1.2
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
webargs==8.2.0
Werkzeug==2.2.3
marshmallow~=3.20.1
python-dotenv~=1.0.0